Allow layers property to be present but empty
This required a small change in the config schema to allow the layers
property to be "null" which is in fact an empty object in the kas file
that is parsed into a NoneType in the code.

Documentation has been updated too, describing the new behaviour when
leaving the layers property empty.

This is not even a breaking change in the configuration file, because
even if documented, leaving the layers property empty or none would
not work anyways before this patch, because according to the config
schema it would have been considered as an error.

Signed-off-by: Giulio Girardi <giulio.girardi@protechgroup.it>